Create sleek non offKilter component
