+++ src/components/LandingView.tsx
@@ -1,33 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
-import { motion } from 'framer-motion';
 import View from './View';
-import getRotation from '../utils/get-rotation';
+import OffKilter from './OffKilter';
 
-const Title = styled(motion.h1)``;
-
-const Subtitle = styled(motion.h2)`
+const Subtitle = styled(OffKilter)`
   font-weight: 200;
 `;
-
-const variants = {
-  normal: {
-    rotate: 0,
-  },
-  offKilter: {
-    rotate: getRotation(1.5, 2),
-    transition: {
-      type: 'spring',
-      bounce: 0.75,
-    },
-  },
-};
 
 const LandingView = () => {
   return (
     <View>
-      <Title variants={variants}>jonathan horn</Title>
-      <Subtitle variants={variants}>full stack web developer</Subtitle>
+      <OffKilter asEl="h1">jonathan horn</OffKilter>
+      <Subtitle asEl="h2">full stack web developer</Subtitle>
     </View>
   );
 };
